// src/lib/server/types/db/runs/purple/purpleParty.ts
export interface PurplePartyRunDbModel {
	RunID: string;
	PartySize: string;
	Patch: string;
	Region: string;
	Rank: string;
	Time: string;
	Notes: string;
	RunServer: string;
	SubmissionTime: string;
	SubmitterID: string;
	ModNotes: string;
	PlayerID: string;
	RunCharacterName: string;
	ShipOverride: string;
	MainClass: string;
	SubClass: string;
	WeaponInfo1: string;
	WeaponInfo2: string;
	WeaponInfo3: string;
	WeaponInfo4: string;
	WeaponInfo5: string;
	WeaponInfo6: string;
	LinkPOV: string;
	PlayerName: string;
	CharacterName: string;
	Description: string;
	Youtube: string;
	Twitch: string;
	Twitter: string;
	Discord: string;
	PreferredName: string;
	Server: string;
	Ship: string;
	Flag: string;
	BackgroundType: string;
	BackgroundImage: string;
	BackgroundColor: string;
	NameType: string;
	NameColor1: string;
	NameColor2: string;
}